requirements.txt
numpy
matplotlib
pandas
scikit-learn
urllib
gzip
